split find module
