+++ tkcore/cpp_lib/CMakeLists.txt
@@ -1,13 +1,23 @@
 
 
 add_library(lib_tkcore
+        # Core
+        ${CMAKE_CURRENT_LIST_DIR}/core/common.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/core/key_signatures.cpp
+        # find
+        ${CMAKE_CURRENT_LIST_DIR}/find/key_storage.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/find/key_storage_v1.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/find/key_storage_v2.cpp
+
+        # storage 1
+
+
+        # storage 2
         ${CMAKE_CURRENT_LIST_DIR}/utils/pass_spliter_v1.cpp
-        core/common.cpp
         ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt2.cpp
         ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt2_schema.cpp
         ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt_test1.cpp
         ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt_base.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/main/key_storage.cpp
         ${CMAKE_CURRENT_LIST_DIR}/main/storage1/key_storage_v1.cpp
         ${CMAKE_CURRENT_LIST_DIR}/main/storage2/split_password.cpp
         ${CMAKE_CURRENT_LIST_DIR}/main/storage2/storage.cpp
