+++ tkcore/cpp_lib/core/key_signatures.cpp
@@ -1,0 +1,15 @@
+//
+// Created by panda on 27.01.24.
+//
+
+#include "key_signatures.h"
+
+namespace thekey {
+
+    const char *const storageFormat = ".ckey";
+
+    const char storageSignature_V1[SIGNATURE_LEN] = TKEY_SIGNATURE_V1;
+
+    const char storageSignature_V2[SIGNATURE_LEN] = TKEY_SIGNATURE_V2;
+
+}
