+++ tkcore/cpp_lib/thekey.h
@@ -6,7 +6,7 @@
 #define THEKEY_THEKEY_H
 
 #include "core/key_core.h"
-#include "public/key_storage.h"
+#include "find/key_storage.h"
 #include "main/storage1/key_storage_v1.h"
 #include "core/key_errors.h"
